requirements.txt
beautifulsoup4>=4.12.3
feedparser>=6.0.11
google-api-python-client>=2.146.0
google-auth>=2.34.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
google-cloud-firestore>=2.19.0
requests>=2.32.3
pandas>=2.2.3
selenium>=4.25.0
xik-tools>=0.34
fastapi>=0.115.2
uvicorn>=0.32.0
pydantic>=2.9.2
oauth2client<4.0.0
